+++ settingsinstance.php
@@ -32,7 +32,7 @@ include "./utils.php";
 
             if (isset($instance_config["general"]) and isset($instance_config["management"]) and isset($instance_config["prerecorded"]) and isset($instance_config["realtime"]) and isset($instance_config["dashcam"]) and isset($instance_config["developer"])) { // If this statement is true, then the configuration is likely vanilla Predator.
             } else if (isset($instance_config["general"]) and isset($instance_config["image"]) and isset($instance_config["alpr"]) and isset($instance_config["network"])) { // If this statement is true, then the configuration is likely Predator Fabric.
-                echo "<p class=\"error\">The current instance configuration file appears to be for Predator Fabric, not vanilla Predator. " . $config["product_name"] . " is designed to be used with full Predator, not modified variants.</p>";
+                echo "<p class=\"error\">The current instance configuration file appears to be for Predator Fabric, not vanilla Predator. " . $config["product_name"] . " is designed to be used only with full Predator, not modified variants.</p>";
                 exit();
             } else { // If neither of the statements above are true, then it is likely that the configuration file is corrupt.
                 echo "<p class=\"error\">The instance configuration appears to be incomplete. Please ensure that your Predator configuration file is valid.</p>";
@@ -47,11 +47,15 @@ include "./utils.php";
 
             $input_values = array();
             $input_values["dashcam"]["saving"]["unsaved_history_length"] = intval($_POST["dashcam>saving>unsaved_history_length"]);
+            $input_values["dashcam"]["saving"]["segment_length"] = floatval($_POST["dashcam>saving>segment_length"]);
             $input_values["dashcam"]["parked"]["enabled"] = $_POST["dashcam>parked>enabled"];
             $input_values["dashcam"]["parked"]["conditions"]["speed"] = intval($_POST["dashcam>parked>conditions>speed"]);
             $input_values["dashcam"]["parked"]["conditions"]["time"] = intval($_POST["dashcam>parked>conditions>time"]);
-            $input_values["dashcam"]["capture"]["resolution"] = $_POST["dashcam>capture>resolution"];
-            $input_values["dashcam"]["capture"]["segment_length"] = floatval($_POST["dashcam>capture>segment_length"]);
+            $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"] = floatval($_POST["dashcam>parked>recording>highlight_motion>enabled"]);
+            $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["color"] = $_POST["dashcam>parked>recording>highlight_motion>color"];
+            $input_values["dashcam"]["parked"]["recording"]["sensitivity"] = floatval($_POST["dashcam>parked>recording>sensitivity"]);
+            $input_values["dashcam"]["parked"]["recording"]["timeout"] = floatval($_POST["dashcam>parked>recording>timeout"]);
+            $input_values["dashcam"]["capture"]["video"]["resolution"] = $_POST["dashcam>capture>video>resolution"];
             $input_values["dashcam"]["stamps"]["size"] = floatval($_POST["dashcam>stamps>size"]);
             $input_values["dashcam"]["stamps"]["main"]["date"]["enabled"] = $_POST["dashcam>stamps>main>date>enabled"];
             $input_values["dashcam"]["stamps"]["main"]["time"]["enabled"] = $_POST["dashcam>stamps>main>time>enabled"];
@@ -61,6 +65,8 @@ include "./utils.php";
             $input_values["dashcam"]["stamps"]["gps"]["altitude"]["enabled"] = $_POST["dashcam>stamps>gps>altitude>enabled"];
             $input_values["dashcam"]["stamps"]["gps"]["speed"]["enabled"] = $_POST["dashcam>stamps>gps>speed>enabled"];
             $input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] = strtolower(strval($_POST["dashcam>stamps>gps>speed>unit"]));
+            $input_values["dashcam"]["capture"]["audio"]["enabled"]= $_POST["dashcam>capture>audio>enabled"];
+            $input_values["dashcam"]["capture"]["audio"]["merge"]= $_POST["dashcam>capture>audio>merge"];
 
 
             // Validate the values from the input form.
@@ -69,17 +75,44 @@ include "./utils.php";
                 $valid = true; // By default, assume the configuration is valid until an invalid value is found.
 
 
+
+                $original_device_count = sizeof($instance_config["dashcam"]["capture"]["video"]["devices"]); // Count the number of capture devices already in the instance configuration.
+                $instance_config["dashcam"]["capture"]["video"]["devices"] = array(); // Reset the list of devices in the loaded instance configuration.
+                for ($i = 0; $i <= $original_device_count + 1; $i++) { // Run once for each device in the configuration, plus one to account for the new entry.
+                    $device_name = $_POST["dashcam>capture>video>devices>" . $i . ">name"]; // This will be the key for the capture device.
+                    $device_index = intval($_POST["dashcam>capture>video>devices>" . $i . ">index"]); // This is the index ID of the capture device.
+                    if (strlen($device_name) > 0) { // Check to see if the device name is set.
+                        if ($device_index >= 0) { // Check to see if the device index if a valid number.
+                            if (!in_array($device_index, $instance_config["dashcam"]["capture"]["video"]["devices"])) { // Check to make sure there is no capture device that already uses this device index.
+                                $instance_config["dashcam"]["capture"]["video"]["devices"][$device_name] = $device_index; // Add this device to the instance configuration.
+                            } else {
+                                echo "<p class='error'>The index for <b>dashcam>capture>video>devices>" . $device_name . "</b> value is already used by another capture device.</p>";
+                                $valid = false;
+                            }
+                        } else {
+                            echo "<p class='error'>The index for <b>dashcam>capture>video>devices>" . $device_name . "</b> value is invalid.</p>";
+                            $valid = false;
+                        }
+                    }
+                }
+
+
+
                 if ($input_values["dashcam"]["saving"]["unsaved_history_length"] < 2) { echo "<p class='error'>The <b>dashcam>saving>unsaved_history_length</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>saving>unsaved_history_length is within the expected range.
 
 
                 if (strtolower($input_values["dashcam"]["parked"]["enabled"]) == "on") { $input_values["dashcam"]["parked"]["enabled"] = true; } else { $input_values["dashcam"]["parked"]["enabled"] = false; } // Convert the dashcam>parked>enabled value to a boolean.
                 if ($input_values["dashcam"]["parked"]["conditions"]["speed"] < 0 or $input_values["dashcam"]["parked"]["conditions"]["speed"] > 10) { echo "<p class='error'>The <b>dashcam>parked>conditions>speed</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>parked>conditions>speed is within the expected range.
-                if ($input_values["dashcam"]["parked"]["conditions"]["time"] < 10 or $input_values["dashcam"]["parked"]["conditions"]["speed"] > 600) { echo "<p class='error'>The <b>dashcam>parked>conditions>speed</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>parked>conditions>time is within the expected range.
+                if ($input_values["dashcam"]["parked"]["conditions"]["time"] < 10 or $input_values["dashcam"]["parked"]["conditions"]["time"] > 600) { echo "<p class='error'>The <b>dashcam>parked>conditions>time</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>parked>conditions>time is within the expected range.
+                if (strtolower($input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"]) == "on") { $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"] = true; } else { $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"] = false; } // Convert the dashcam>parked>recording>highlight_motion>enabled value to a boolean.
+                $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["color"] = sscanf($input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["color"], "#%02x%02x%02x");
+                if ($input_values["dashcam"]["parked"]["recording"]["sensitivity"] < 0.001 or $input_values["dashcam"]["parked"]["recording"]["sensitivity"] > 0.8) { echo "<p class='error'>The <b>dashcam>parked>recording>sensitivity</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>parked>recording>sensitivity is within the expected range.
+                if ($input_values["dashcam"]["parked"]["recording"]["timeout"] < 0 or $input_values["dashcam"]["parked"]["recording"]["timeout"] > 60) { echo "<p class='error'>The <b>dashcam>parked>recording>timeout</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>parked>recording>timeout is within the expected range.
 
 
-                if ($input_values["dashcam"]["capture"]["resolution"] !== "426x240" and $input_values["dashcam"]["capture"]["resolution"] !== "640x360" and $input_values["dashcam"]["capture"]["resolution"] !== "640x480" and $input_values["dashcam"]["capture"]["resolution"] !== "960x540" and $input_values["dashcam"]["capture"]["resolution"] !== "1280x720" and $input_values["dashcam"]["capture"]["resolution"] !== "1920x1080" and $input_values["dashcam"]["capture"]["resolution"] !== "2560x1440" and $input_values["dashcam"]["capture"]["resolution"] !== "3840x2160" and $input_values["dashcam"]["capture"]["resolution"] !== "7680x4320") { echo "<p class='error'>The <b>dashcam>capture>resolution</b> value is not a recognized option.</p>"; $valid = false; } // Validate that the dashcam>capture>resolution is an expected option.
+                if ($input_values["dashcam"]["capture"]["video"]["resolution"] !== "426x240" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "640x360" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "640x480" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "960x540" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "1280x720" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "1920x1080" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "2560x1440" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "3840x2160" and $input_values["dashcam"]["capture"]["video"]["resolution"] !== "7680x4320") { echo "<p class='error'>The <b>dashcam>capture>video>resolution</b> value is not a recognized option.</p>"; $valid = false; } // Validate that the dashcam>capture>video>resolution is an expected option.
 
-                if ($input_values["dashcam"]["capture"]["segment_length"] < 0 or $input_values["dashcam"]["capture"]["segment_length"] > 360) { echo "<p class='error'>The <b>dashcam>capture>segment_length</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>capture>segment_length value is within the expected range.
+                if ($input_values["dashcam"]["saving"]["segment_length"] < 0 or $input_values["dashcam"]["saving"]["segment_length"] > 360) { echo "<p class='error'>The <b>dashcam>saving>segment_length</b> value is invalid.</p>"; $valid = false; } // Validate that the dashcam>saving>segment_length value is within the expected range.
 
 
 
@@ -100,15 +133,23 @@ include "./utils.php";
                 if ($input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] !== "mph" and $input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] !== "kph" and $input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] !== "mps" and $input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] !== "fps" and $input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"] !== "knot") { echo "<p class='error'>The <b>dashcam>stamps>gps>speed>unit</b> is not an expected value.</p>"; $valid = false; }
 
 
+                if (strtolower($input_values["dashcam"]["capture"]["audio"]["enabled"]) == "on") { $input_values["dashcam"]["capture"]["audio"]["enabled"] = true; } else { $input_values["dashcam"]["capture"]["audio"]["enabled"] = false; } // Convert the dashcam>capture>audio>enabled value to a boolean.
+                if (strtolower($input_values["dashcam"]["capture"]["audio"]["merge"]) == "on") { $input_values["dashcam"]["capture"]["audio"]["merge"] = true; } else { $input_values["dashcam"]["capture"]["audio"]["merge"] = false; } // Convert the dashcam>capture>audio>merge value to a boolean.
+
+
                 // Update the instance configuration file.
                 if ($valid == true) { // Check to see if all configuration values were validated.
                     $instance_config["dashcam"]["saving"]["unsaved_history_length"] = intval($input_values["dashcam"]["saving"]["unsaved_history_length"]);
+                    $instance_config["dashcam"]["saving"]["segment_length"] = $input_values["dashcam"]["saving"]["segment_length"];
                     $instance_config["dashcam"]["parked"]["enabled"] = $input_values["dashcam"]["parked"]["enabled"];
                     $instance_config["dashcam"]["parked"]["conditions"]["speed"] = floatval($input_values["dashcam"]["parked"]["conditions"]["speed"]);
                     $instance_config["dashcam"]["parked"]["conditions"]["time"] = floatval($input_values["dashcam"]["parked"]["conditions"]["time"]);
-                    $instance_config["dashcam"]["capture"]["resolution"]["width"] = floatval(explode("x", $input_values["dashcam"]["capture"]["resolution"])[0]);
-                    $instance_config["dashcam"]["capture"]["resolution"]["height"] = floatval(explode("x", $input_values["dashcam"]["capture"]["resolution"])[1]);
-                    $instance_config["dashcam"]["capture"]["segment_length"] = $input_values["dashcam"]["capture"]["segment_length"];
+                    $instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"] = $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"];
+                    $instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["color"] = $input_values["dashcam"]["parked"]["recording"]["highlight_motion"]["color"];
+                    $instance_config["dashcam"]["parked"]["recording"]["sensitivity"] = floatval($input_values["dashcam"]["parked"]["recording"]["sensitivity"]);
+                    $instance_config["dashcam"]["parked"]["recording"]["timeout"] = floatval($input_values["dashcam"]["parked"]["recording"]["timeout"]);
+                    $instance_config["dashcam"]["capture"]["video"]["resolution"]["width"] = floatval(explode("x", $input_values["dashcam"]["capture"]["video"]["resolution"])[0]);
+                    $instance_config["dashcam"]["capture"]["video"]["resolution"]["height"] = floatval(explode("x", $input_values["dashcam"]["capture"]["video"]["resolution"])[1]);
                     $instance_config["dashcam"]["stamps"]["size"] = $input_values["dashcam"]["stamps"]["size"];
                     $instance_config["dashcam"]["stamps"]["main"]["date"]["enabled"] = $input_values["dashcam"]["stamps"]["main"]["date"]["enabled"];
                     $instance_config["dashcam"]["stamps"]["main"]["time"]["enabled"] = $input_values["dashcam"]["stamps"]["main"]["time"]["enabled"];
@@ -118,6 +159,8 @@ include "./utils.php";
                     $instance_config["dashcam"]["stamps"]["gps"]["altitude"]["enabled"] = $input_values["dashcam"]["stamps"]["gps"]["altitude"]["enabled"];
                     $instance_config["dashcam"]["stamps"]["gps"]["speed"]["enabled"] = $input_values["dashcam"]["stamps"]["gps"]["speed"]["enabled"];
                     $instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] = strtolower($input_values["dashcam"]["stamps"]["gps"]["speed"]["unit"]);
+                    $instance_config["dashcam"]["capture"]["audio"]["enabled"] = $input_values["dashcam"]["capture"]["audio"]["enabled"];
+                    $instance_config["dashcam"]["capture"]["audio"]["merge"] = $input_values["dashcam"]["capture"]["audio"]["merge"];
 
                     if (json_encode($instance_config) == true) { // Verify that the data to be saved to the instance configuration file is valid.
                         $instance_configuration_file = $config["instance_directory"] . "/config.json";
@@ -132,84 +175,119 @@ include "./utils.php";
                     }
                 } else { // If one or more configuration values is invalid, then don't update the configuration.
                     echo "<p class='error'>The configuration was not updated.<p>";
+                    $instance_config = load_instance_config($config); // Reload the instance configuration so invalid information is not placed into the input fields.
                 }
             }
 
             ?>
             <form method="post">
                 <div class="buffer">
-                    <h3>Video</h3>
+                    <h3>Saving</h3>
+                    <label for="dashcam>saving>segment_length" title="The length of a video segment, in seconds, before another segment is started.">Segment Length: </label><input type="number" class="compactinput" id="dashcam>saving>segment_length" name="dashcam>saving>segment_length" step="1" min="0" max="3600" value="<?php echo $instance_config["dashcam"]["saving"]["segment_length"]; ?>"> seconds<br><br>
+                    <label for="dashcam>saving>unsaved_history_length" title="The number of segments that can be recorded before unsaved segments start to be overwritten.">History Length: </label><input type="number" class="compactinput" id="dashcam>saving>unsaved_history_length" name="dashcam>saving>unsaved_history_length" step="1" min="2" max="10000" value="<?php echo $instance_config["dashcam"]["saving"]["unsaved_history_length"]; ?>"> segments
+                </div>
+                <div class="buffer">
+                    <h3>Capture</h3>
                     <div class="buffer">
-                        <h4>Capture</h4>
-                        <label for="dashcam>capture>resolution" title="The resolution at which Predator should capture video.">Resolution:</label>
-                        <select id="dashcam>capture>resolution" name="dashcam>capture>resolution">
-                            <option value="426x240" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 240) { echo "selected"; } ?>>240p</option>
-                            <option value="640x360" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 360) { echo "selected"; } ?>>360p</option>
-                            <option value="640x480" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 480) { echo "selected"; } ?>>480p</option>
-                            <option value="960x540" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 540) { echo "selected"; } ?>>540p</option>
-                            <option value="1280x720" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 720) { echo "selected"; } ?>>720p</option>
-                            <option value="1920x1080" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 1080) { echo "selected"; } ?>>1080p</option>
-                            <option value="2560x1440" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 1440) { echo "selected"; } ?>>1440p</option>
-                            <option value="3840x2160" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 2160) { echo "selected"; } ?>>2160p</option>
-                            <option value="7680x4320" <?php if (intval($instance_config["dashcam"]["capture"]["resolution"]["height"]) == 4320) { echo "selected"; } ?>>4320p</option>
+                        <h4>Video</h4>
+                        <label for="dashcam>capture>video>resolution" title="The resolution at which Predator should capture video.">Resolution:</label>
+                        <select id="dashcam>capture>video>resolution" name="dashcam>capture>video>resolution">
+                            <option value="426x240" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 240) { echo "selected"; } ?>>240p</option>
+                            <option value="640x360" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 360) { echo "selected"; } ?>>360p</option>
+                            <option value="640x480" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 480) { echo "selected"; } ?>>480p</option>
+                            <option value="960x540" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 540) { echo "selected"; } ?>>540p</option>
+                            <option value="1280x720" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 720) { echo "selected"; } ?>>720p</option>
+                            <option value="1920x1080" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 1080) { echo "selected"; } ?>>1080p</option>
+                            <option value="2560x1440" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 1440) { echo "selected"; } ?>>1440p</option>
+                            <option value="3840x2160" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 2160) { echo "selected"; } ?>>2160p</option>
+                            <option value="7680x4320" <?php if (intval($instance_config["dashcam"]["capture"]["video"]["resolution"]["height"]) == 4320) { echo "selected"; } ?>>4320p</option>
                         </select><br><br>
-                    </div>
-                    <div class="buffer">
-                        <h4>Saving</h4>
-                        <label for="dashcam>capture>segment_length" title="The length of a video segment, in seconds, before another segment is started.">Segment Length: </label><input type="number" class="compactinput" id="dashcam>capture>segment_length" name="dashcam>capture>segment_length" step="1" min="0" max="3600" value="<?php echo $instance_config["dashcam"]["capture"]["segment_length"]; ?>"> seconds<br><br>
-                        <label for="dashcam>saving>unsaved_history_length" title="The number of segments that can be recorded before unsaved segments start to be overwritten.">Unsaved History Length: </label><input type="number" class="compactinput" id="dashcam>saving>unsaved_history_length" name="dashcam>saving>unsaved_history_length" step="1" min="2" max="10000" value="<?php echo $instance_config["dashcam"]["saving"]["unsaved_history_length"]; ?>"> segments
-                    </div>
-                    <div class="buffer">
-                        <h4>Parking</h4>
-                        <label for="dashcam>parked>enabled" title="Determines whether or not Predator will go into parking mode when the vehicle is stopped for extended periods of time.">Enabled: </label><input type="checkbox" id="dashcam>parked>enabled" name="dashcam>parked>enabled" <?php if ($instance_config["dashcam"]["parked"]["enabled"] == true) { echo "checked"; } ?>><br><br>
                         <div class="buffer">
-                            <h5>Conditions</h5>
-                            <label for="dashcam>parked>conditions>speed" title="The speed below which Predator considers the vehicle to be stopped.">Speed: </label><input type="number" class="compactinput" id="dashcam>parked>conditions>speed" name="dashcam>parked>conditions>speed" step="0.1" min="0" max="5" value="<?php echo $instance_config["dashcam"]["parked"]["conditions"]["speed"]; ?>"> m/s<br><br>
-                            <label for="dashcam>parked>conditions>time" title="The length of time the vehicle needs to be stopped before entering into parking mode.">Time: </label><input type="number" class="compactinput" id="dashcam>parked>conditions>time" name="dashcam>parked>conditions>time" step="10" min="10" max="600" value="<?php echo $instance_config["dashcam"]["parked"]["conditions"]["time"]; ?>"> seconds<br><br>
-                        </div>
-                    </div>
-                    <div class="buffer">
-                        <h4>Overlays</h4>
-                        <label for="dashcam>stamps>size" title="The font size of all overlay stamps.">Size: </label><input type="number" class="compactinput" id="dashcam>stamps>size" name="dashcam>stamps>size" step="0.1" min="0.2" max="3" value="<?php echo $instance_config["dashcam"]["stamps"]["size"]; ?>"><br><br>
-                        <div class="buffer">
-                            <h5>Main</h5>
-                            <label for="dashcam>stamps>main>date>enabled">Date: </label><input type="checkbox" id="dashcam>stamps>main>date>enabled" name="dashcam>stamps>main>date>enabled" <?php if ($instance_config["dashcam"]["stamps"]["main"]["date"]["enabled"] == true) { echo "checked"; } ?>><br><br>
-                            <label for="dashcam>stamps>main>time>enabled">Time: </label><input type="checkbox" id="dashcam>stamps>main>time>enabled" name="dashcam>stamps>main>time>enabled" <?php if ($instance_config["dashcam"]["stamps"]["main"]["time"]["enabled"] == true) { echo "checked"; } ?>><br><br>
-                            <label for="dashcam>stamps>main>message_1">License Plate: </label><input type="text" id="dashcam>stamps>main>message_1" name="dashcam>stamps>main>message_1" pattern="[a-zA-Z0-9-_ ]{0,12}" value="<?php echo $instance_config["dashcam"]["stamps"]["main"]["message_1"]; ?>"><br><br>
+                            <h4>Cameras</h4>
                             <?php
-                            if (pro_flat()) {
-                                if ($config["\160\x72\157\x64\165\143\x74\137\156\141\x6d\145"] == "\117\x70\164\151\143\40\x50\x72\x6f") {
-                                    echo '<label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" name="dashcam>stamps>main>message_2" pattern="[a-zA-Z0-9-_ ]{0,50}" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '"><br><br>';
-                                } else {
-                                    echo '<div style="color:#aaaaaa;"><label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '" disabled style="color:inherit"></div><br><br>';
-                                }
-                            } else {
-                                echo '<div style="color:#aaaaaa;"><label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '" disabled style="color:inherit"></div>(Optic Pro only)<br><br>';
+                            $displayed_cameras = 0;
+                            foreach (array_keys($instance_config["dashcam"]["capture"]["video"]["devices"]) as $key) {
+                                echo '<div class="buffer">';
+                                echo '    <h5>Device "' . $key . '"</h5>';
+                                echo '    <label for="dashcam>capture>video>devices>' . $displayed_cameras . '>name" title="The name that will be used as this capture device\'s ID.">Name: </label><input type="text" class="compactinput" id="dashcam>capture>video>devices>' . $displayed_cameras . '>name" name="dashcam>capture>video>devices>' . $displayed_cameras . '>name" min="0" max="10" value="' . $key . '"><br><br>';
+                                echo '    <label for="dashcam>capture>video>devices>' . $displayed_cameras . '>index" title="The index number of the capture device on the system.">Index: </label><input type="number" class="compactinput" id="dashcam>capture>video>devices>' . $displayed_cameras . '>index" name="dashcam>capture>video>devices>' . $displayed_cameras . '>index" step="1" min="0" max="10" value="' . $instance_config["dashcam"]["capture"]["video"]["devices"][$key] . '">';
+                                echo '</div>';
+                                $displayed_cameras++;
                             }
                             ?>
+                            <div class="buffer">
+                                <h5>New Device</h5>
+                                <label for="dashcam>capture>video>devices><?php echo $displayed_cameras; ?>>name" title="The name that will be used as this capture device's ID.">Name: </label><input type="text" class="compactinput" id="dashcam>capture>video>devices><?php echo $displayed_cameras; ?>>name" name="dashcam>capture>video>devices><?php echo $displayed_cameras; ?>>name" max="10"><br><br>
+                                <label for="dashcam>capture>video>devices<?php echo $displayed_cameras; ?>>index" title="The index number of the capture device on the system.">Index: </label><input type="text" class="compactinput" id="dashcam>capture>video>devices><?php echo $displayed_cameras; ?>>index" name="dashcam>capture>video>devices><?php echo $displayed_cameras; ?>>index" step="1" min="0" max="10">
+                            </div>
+                        </div>
+                    </div>
+                    <div class="buffer">
+                        <h4>Audio</h4>
+                        <label for="dashcam>capture>audio>enabled" title="Determines audio will be recorded during video capture.">Enabled: </label><input type="checkbox" id="dashcam>capture>audio>enabled" name="dashcam>capture>audio>enabled" <?php if ($instance_config["dashcam"]["capture"]["audio"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                        <label for="dashcam>capture>audio>merge" title="Determines if the separate audio and video file will be merged at the end of each segment.">Merge: </label><input type="checkbox" id="dashcam>capture>audio>merge" name="dashcam>capture>audio>merge" <?php if ($instance_config["dashcam"]["capture"]["audio"]["merge"] == true) { echo "checked"; } ?>>
+                    </div>
+                </div>
+                <div class="buffer">
+                    <h3>Parking</h3>
+                    <label for="dashcam>parked>enabled" title="Determines whether or not Predator will go into parking mode when the vehicle is stopped for extended periods of time.">Enabled: </label><input type="checkbox" id="dashcam>parked>enabled" name="dashcam>parked>enabled" <?php if ($instance_config["dashcam"]["parked"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                    <div class="buffer">
+                        <h4>Conditions</h4>
+                        <label for="dashcam>parked>conditions>speed" title="The speed below which Predator considers the vehicle to be stopped.">Speed: </label><input type="number" class="compactinput" id="dashcam>parked>conditions>speed" name="dashcam>parked>conditions>speed" step="0.1" min="0" max="5" value="<?php echo $instance_config["dashcam"]["parked"]["conditions"]["speed"]; ?>"> m/s<br><br>
+                        <label for="dashcam>parked>conditions>time" title="The length of time the vehicle needs to be stopped before entering into parking mode.">Time: </label><input type="number" class="compactinput" id="dashcam>parked>conditions>time" name="dashcam>parked>conditions>time" step="10" min="10" max="600" value="<?php echo $instance_config["dashcam"]["parked"]["conditions"]["time"]; ?>"> seconds
+                    </div>
+                    <div class="buffer">
+                        <h4>Recording</h4>
+                        <div class="buffer">
+                            <h5>Highlight Motion</h5>
+                            <label for="dashcam>parked>recording>highlight_motion>enabled" title="Determines if bounding boxes will be drawn around detected movement while parked.">Enabled: </label><input type="checkbox" id="dashcam>parked>recording>highlight_motion>enabled" name="dashcam>parked>recording>highlight_motion>enabled" <?php if ($instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                            <label for="dashcam>parked>recording>highlight_motion>color" title="The color that motion bounding boxes will be drawn in.">Color: </label><input type="color" id="dashcam>parked>recording>highlight_motion>color" name="dashcam>parked>recording>highlight_motion>color" value="<?php echo sprintf("#%02x%02x%02x", $instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["color"][0], $instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["color"][1], $instance_config["dashcam"]["parked"]["recording"]["highlight_motion"]["color"][2]) ?>">
+                        </div>
+                        <label for="dashcam>parked>recording>sensitivity" title="The minimum fraction of the frame that needs to be in motion to trigger recording.">Sensitivity: </label><input type="number" class="compactinput" id="dashcam>parked>recording>sensitivity" name="dashcam>parked>recording>sensitivity" step="0.001" min="0.001" max="1" value="<?php echo $instance_config["dashcam"]["parked"]["recording"]["sensitivity"]; ?>"><br><br>
+                        <label for="dashcam>parked>recording>timeout" title="The length of time that video will be recorded after motion is last detected.">Timeout: </label><input type="number" class="compactinput" id="dashcam>parked>recording>timeout" name="dashcam>parked>recording>timeout" step="1" min="0" max="60" value="<?php echo $instance_config["dashcam"]["parked"]["recording"]["timeout"]; ?>"> seconds
+                    </div>
+                </div>
+                <div class="buffer">
+                    <h3>Overlays</h3>
+                    <label for="dashcam>stamps>size" title="The font size of all overlay stamps.">Size: </label><input type="number" class="compactinput" id="dashcam>stamps>size" name="dashcam>stamps>size" step="0.1" min="0.2" max="3" value="<?php echo $instance_config["dashcam"]["stamps"]["size"]; ?>"><br><br>
+                    <div class="buffer">
+                        <h4>Main</h4>
+                        <label for="dashcam>stamps>main>date>enabled">Date: </label><input type="checkbox" id="dashcam>stamps>main>date>enabled" name="dashcam>stamps>main>date>enabled" <?php if ($instance_config["dashcam"]["stamps"]["main"]["date"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                        <label for="dashcam>stamps>main>time>enabled">Time: </label><input type="checkbox" id="dashcam>stamps>main>time>enabled" name="dashcam>stamps>main>time>enabled" <?php if ($instance_config["dashcam"]["stamps"]["main"]["time"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                        <label for="dashcam>stamps>main>message_1">License Plate: </label><input type="text" id="dashcam>stamps>main>message_1" name="dashcam>stamps>main>message_1" pattern="[a-zA-Z0-9-_ ]{0,12}" value="<?php echo $instance_config["dashcam"]["stamps"]["main"]["message_1"]; ?>"><br><br>
+                        <?php
+                        if (pro_flat()) {
+                            if ($config["\160\x72\157\x64\165\143\x74\137\156\141\x6d\145"] == "\117\x70\164\151\143\40\x50\x72\x6f") {
+                                echo '<label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" name="dashcam>stamps>main>message_2" pattern="[a-zA-Z0-9-_ ]{0,50}" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '">';
+                            } else {
+                                echo '<div style="color:#aaaaaa;"><label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '" disabled style="color:inherit"></div>';
+                            }
+                        } else {
+                            echo '<div style="color:#aaaaaa;"><label for="dashcam>stamps>main>message_2">Custom Message: </label><input type="text" id="dashcam>stamps>main>message_2" value="' . $instance_config["dashcam"]["stamps"]["main"]["message_2"] . '" disabled style="color:inherit"></div>(Optic Pro only)';
+                        }
+                        ?>
+                    </div>
+                    <div class="buffer">
+                        <h4>GPS</h4>
+                        <div class="buffer">
+                            <h5>Location</h5>
+                            <label for="dashcam>stamps>gps>location>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>location>enabled" name="dashcam>stamps>gps>location>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["location"]["enabled"] == true) { echo "checked"; } ?>>
                         </div>
                         <div class="buffer">
-                            <h5>GPS</h5>
-                            <div class="buffer">
-                                <h6>Location</h6>
-                                <label for="dashcam>stamps>gps>location>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>location>enabled" name="dashcam>stamps>gps>location>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["location"]["enabled"] == true) { echo "checked"; } ?>><br><br>
-                            </div>
-                            <div class="buffer">
-                                <h6>Altitude</h6>
-                                <label for="dashcam>stamps>gps>altitude>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>altitude>enabled" name="dashcam>stamps>gps>altitude>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["altitude"]["enabled"] == true) { echo "checked"; } ?>><br><br>
-                            </div>
-                            <div class="buffer">
-                                <h6>Speed</h6>
-                                <label for="dashcam>stamps>gps>speed>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>speed>enabled" name="dashcam>stamps>gps>speed>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["enabled"] == true) { echo "checked"; } ?>><br><br>
-                                <label for="dashcam>stamps>gps>speed>unit" title="The unit the speed stamp is displayed in.">Units:</label>
-                                <select id="dashcam>stamps>gps>speed>unit" name="dashcam>stamps>gps>speed>unit">
-                                    <option value="mph" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "mph") { echo "selected"; } ?>>miles per hour</option>
-                                    <option value="kph" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "kph") { echo "selected"; } ?>>kilometers per hour</option>
-                                    <option value="mps" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "mps") { echo "selected"; } ?>>meters per second</option>
-                                    <option value="fps" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "fps") { echo "selected"; } ?>>feet per second</option>
-                                    <option value="knot" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "knot") { echo "selected"; } ?>>knots</option>
-                                </select><br><br>
-                            </div>
+                            <h5>Altitude</h5>
+                            <label for="dashcam>stamps>gps>altitude>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>altitude>enabled" name="dashcam>stamps>gps>altitude>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["altitude"]["enabled"] == true) { echo "checked"; } ?>>
+                        </div>
+                        <div class="buffer">
+                            <h5>Speed</h5>
+                            <label for="dashcam>stamps>gps>speed>enabled">Enabled: </label><input type="checkbox" id="dashcam>stamps>gps>speed>enabled" name="dashcam>stamps>gps>speed>enabled" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["enabled"] == true) { echo "checked"; } ?>><br><br>
+                            <label for="dashcam>stamps>gps>speed>unit" title="The unit the speed stamp is displayed in.">Units:</label>
+                            <select id="dashcam>stamps>gps>speed>unit" name="dashcam>stamps>gps>speed>unit">
+                                <option value="mph" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "mph") { echo "selected"; } ?>>miles per hour</option>
+                                <option value="kph" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "kph") { echo "selected"; } ?>>kilometers per hour</option>
+                                <option value="mps" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "mps") { echo "selected"; } ?>>meters per second</option>
+                                <option value="fps" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "fps") { echo "selected"; } ?>>feet per second</option>
+                                <option value="knot" <?php if ($instance_config["dashcam"]["stamps"]["gps"]["speed"]["unit"] == "knot") { echo "selected"; } ?>>knots</option>
+                            </select>
                         </div>
                     </div>
                 </div>