+++ CHANGELOG.md
@@ -65,10 +65,11 @@ January 15th, 2024
 *Release date to be determined*
 
 - Updated instance configuration page.
-    - Removed the frame-rate configuration option to reflect Predator's updated configuration layout.
+    - Updated the configuration interface to reflect Predator V10's configuration layout.
     - Added more resolution options.
     - Added more GPS video overlay stamp configuration options.
     - Added parking mode configuration options.
+    - Added the ability to add, remove, and modify capture devices.
 - Updated the storage page to be compatible with Predator V10.0.
     - Optic is now compatible with the new Predator configuration layout.
     - Optic now recognizes separate audio files and displays them next to their associated video file.